==> 04_pandas_dados/02_importando_dados.ipynb <==
# Relatório de análise 1

Recebemos uma base de dados com valores de aluguel de imóveis no Rio de Janeiro. O trabalho será verificar alguns padrões usando esta base.

## Importando a base de dados

import pandas as pd

# importando a base de dados
dados = pd.read_csv('dados/aluguel.csv', sep = ';')
dados

type(dados) # retorna falando que é um objeto do tipo DataFrame

dados.info() # retorna as características das colunas do DF

dados.head() # mostra as 5 primeiras linhas

## Informações gerais sobre a base de dados

dados.dtypes # object no pandas é um string

tipos_de_dados = pd.DataFrame(dados.dtypes, columns = ['Tipos de dados'])

tipos_de_dados.columns.name = 'Variáveis'
tipos_de_dados

# linhas e variáveis
print(f'A base de dados apresenta {dados.shape[0]} registros (imóveis) e {dados.shape[1]} colunas.')

Nesta aula, sobre pandas, aprendemos:

 * Como importar a biblioteca (`import pandas as pd`)
 * Como ler fontes de dados diferentes:
    * Uma base CSV (`pd.read_csv(...)`)
    * Uma base JSON (`pd.read_json(...)`)
    * Uma base TXT (`pd.read_table(...)`)
    * Um arquivo EXCEL (`pd.read_excel(...)`)
    * Uma página HTML (`pd.read_html(...)`)
* Vários métodos e atributos úteis de dataframes, como:
    * `info()`
    * `head()`
    * `dtypes`
    * `columns`
    * `shape`

E sobre Jupyter, vimos como:

 * Criar diferentes tipos de células dentro do Jupyter
 * Acessar a documentação
 * Como reexecutar todas as células
