==> 04_pandas_dados/03_series_index.ipynb <==
# Relatório de análise II

## Tipos de imóveis

Como para a seguradora alguns tipos de imóveis não são interessantes para a análise (ela vai trabalhar apenas com imóveis residenciais), precisamos selecionar apenas os tipos de imóveis de interesse.

import pandas as pd

dados = pd.read_csv('dados/aluguel.csv', sep = ';')

dados.head()

dados['Tipo']

tipo_de_imovel = dados['Tipo']

type(tipo_de_imovel)

tipo_de_imovel.drop_duplicates(inplace = True) # inplace força que a alteração aconteça com a própria variável

tipo_de_imovel

## Organizando a visualização

tipo_de_imovel = pd.DataFrame(tipo_de_imovel)
tipo_de_imovel

tipo_de_imovel.index = range(len(tipo_de_imovel))
tipo_de_imovel

tipo_de_imovel.columns.name = 'Id'

tipo_de_imovel

Nesta aula, aprendemos:

 * Como selecionar uma variável do dataframe (por exemplo, `dados['Tipo']` ou `dados.Tipo`)
 * Que um dataframe é composto de vários Series
 * Como eliminar duplicatas (pelo método `drop_duplicates()`)
 * Como redefinir o index de um dataframe e series (atributo `index`)
 * Como concatenar dataframes (`df.concat(dfs...)`, lembrando do `axis`)
 * Como criar novos dataframes baseados em estruturas de dados Python (lista, dicionários ou tuples)
