==> 04_pandas_dados/04_filtrando_dados.ipynb <==
# Relatório de análise III

Filtrando os imóveis residenciais para continuar com a análise

## Imóveis residenciais

import pandas as pd

dados = pd.read_csv('dados/aluguel.csv', sep = ';')
dados.head()

# verificando os tipos de imóveis
list(dados['Tipo'].drop_duplicates())

# determinando quais tipos são residencial - deixei comentado o que na aula não foi considerado como residencial
residencial = ['Quitinete',
 'Casa',
 'Apartamento',
 'Casa de Condomínio',
 #'Flat',
 'Casa de Vila',
 #'Loft',
 #'Studio'
              ]
residencial

selecao = dados['Tipo'].isin(residencial) # preparando a condição para filtrar

dados_residencial = dados[selecao]
dados_residencial

# verificando se a selecao funcionou
list(dados_residencial['Tipo'].drop_duplicates())

# atualizando o index
dados_residencial.index = range(len(dados_residencial))
dados_residencial

## Exportando a base de dados

dados_residencial.to_csv('dados/aluguel_residencial.csv', sep = ';')

# testando o arquivo salvo
df_residencial = pd.read_csv('dados/aluguel_residencial.csv', sep = ';')
df_residencial
# note que como quando salvamos não colocamos que não era para o índice ser exportado

dados_residencial.to_csv('dados/aluguel_residencial.csv', sep = ';', index = False)
# precisamos incluir index = False para que o índice não seja exportado

# testando novamente o arquivo
df_residencial = pd.read_csv('dados/aluguel_residencial.csv', sep = ';')
df_residencial

Nesta aula, vimos como:

 * Criar uma Series booleana usando o método `isin(..)` a partir do dataframe
 * Filtrar os dados de um dataframe baseado na Series booleana
 * Exportar e gravar os dados do dataframe (método `to_csv()`)
 * Ordenar os dados de um dataframe (métodos `sort_values()` e `sort_index()`)
