==> 04_pandas_dados/05_frequencias_imoveis.ipynb <==
# Relatório de análise IV

Foi passada a orientação para identificar:
 * Os imóveis classificados com tipo 'Apartamento'.
 * Os imóveis classificados com tipos 'Casa', 'Casa de Condomínio' e 'Casa de Vila'.
 * Os imóveis com área entre 60 e 100 metros quadrados, incluindo os limites.
 * Os imóveis que tenham pelo menos 4 quartos e aluguel menor que R$ 2.000,00.

## Seleções e Frequências

import pandas as pd

df = pd.read_csv('dados/aluguel_residencial.csv', sep = ';')
df.head()

# imóveis com tipo 'Apartamento'
selecao = df['Tipo'] == 'Apartamento'
n1 = len(df[selecao]) # obtendo a frequencia, numero de imoveis nesta classificação
n1

# imóveis com tipo 'Casa', 'Casa de Vila', 'Casa de Condomínio'
selecao = df['Tipo'].isin(['Casa', 'Casa de Vila', 'Casa de Condomínio'])
n2 = len(df[selecao])
n2

# imóveis com área entre 60 e 100 m2
selecao = df['Area'].isin(range(60,101))
n3 = len(df[selecao])
n3

# imóveis com pelo menos 4 quartos E aluguel menor que 2000 reais
selecao = (df['Quartos'] >= 4) & (df['Valor'] < 2000)
n4 = len(df[selecao])
n4

print(f"São {n1} imóveis classificados com tipo 'Apartamento'.")
print(f"São {n2} imóveis classificados com tipos 'Casa', 'Casa de Condomínio' e 'Casa de Vila'.")
print(f"São {n3} imóveis com área entre 60 e 100 metros quadrados, incluindo os limites.")
print(f"São {n4} imóveis que têm pelo menos 4 quartos e aluguel menor que R$ 2.000,00.")

Nesta aula, aprendemos:

 * Formas de seleção e frequências
 * Seleção com a condição `OR` (|)
 * Seleção com a condição `AND` (&)
 * Como criar um Index com `split()`
 * Seleção por linha e coluna em um dataframe:
     * Utilizando os índices numéricos (`iloc`)
     * e os rótulos das linhas (`loc`)
