==> 04_pandas_dados/06_dados_faltantes.ipynb <==
# Relatório de análise V

## Tratamento de dados faltantes

import pandas as pd

df = pd.read_csv('dados/aluguel_residencial.csv', sep = ';')
df.head()

Identificando valores null

df.isnull()

df.info()
# dá informação da quantidade de elementos não nulos de cada variável

Podemos ver que as variáveis 'Valor', 'Condomínio' e 'IPTU' tem valores nulos.

df[df['Valor'].isnull()]

Como é a variável que mais importa, vamos remover as linhas que não tem o registro do valor de aluguel.

df.dropna(subset = ['Valor'], inplace = True)

## Tratamento de condicional

O valor do condomínio também é importante para a análise. Então vamos avaliar os valores nulos desta variável.

len(df[df['Condominio'].isnull()])

selecao = (df['Tipo'] == 'Apartamento') & (df['Condominio'].isnull())
df[selecao]

Não faz sentido existir apartamentos que não tenham valor de condomínio. Então vamos eliminar esse caso.

df = df[~selecao]

len(df[df['Condominio'].isnull()])

Um caso ignorado na análise são as casas de condomínio. Vamos dar uma olhada nelas.

selecao = (df['Tipo'] == 'Casa de Condomínio') & (df['Condominio'].isnull())
df[selecao]

Para evitar incompatibilidades com os exercícios futuros, não vou eliminar estes casos. Mas eles se enquadram na mesma questão que o apartamento sem condomínio.

Entende-se que os casos restantes no DataFrame que ainda são Null, na verdade, são valores 0. Então vamos substituí-los.

df = df.fillna({'Condominio': 0, 'IPTU': 0})
df

len(df[df['Condominio'].isnull()])

# salvando em um novo arquivo
df.to_csv('dados/aluguel_residencial_limpo.csv', sep = ';', index = False)

Nesta aula, aprendemos:

 * Tratamento de dados faltantes
 * Como identificar valores nulos (missing values)
 * O método `isnull()` indica se os valores são nulos
 * O método `notnull()` retorna o contrário do método `isnull()`
 * O método `info()` também é uma forma de se verificar a presença de valores faltantes
 * Como remover valores nulos com o método `dropna()`
 * Tratamento condicional
 * Inversão de valores booleanos com `~`
 * Como substituir os missing values com o método `fillna()`
 * Métodos de interpolação: `ffill`, `bfill` e `mean()`
