==> 04_pandas_dados/07_novas_variaveis.ipynb <==
# Relatório de análise VI

import pandas as pd

## Criando novas variáveis

dados = pd.read_csv('dados/aluguel_residencial_limpo.csv', sep = ';')
dados.head()

dados['Valor_m2'] = dados['Valor'] / dados['Area']
dados

# arredondando a variavel calculada
dados['Valor_m2'] = dados['Valor_m2'].round(2)
dados

casa = ['Casa', 'Casa de Condomínio', 'Casa de Vila']
casa

## Agregando variável

# criando uma nova variável que agrega os tipos de imóveis em casa ou apto
dados['Tipo Agregado'] = dados['Tipo'].apply(lambda x: 'Casa' if x in casa else 'Apartamento')
dados

## Excluindo variáveis

dados_aux = pd.DataFrame(dados[['Tipo Agregado', 'Valor_m2', 'Tipo', 'Suites']])
dados_aux

del dados_aux['Tipo']
dados_aux

dados_aux.pop('Suites')

dados_aux

# também é possível usar dados.drop(lista de variáveis)

dados.to_csv('dados/aluguel_residencial_tratado.csv', sep = ';', index = False)
